misc update

+++ src/insert_sort.cc
@@ -3,7 +3,7 @@
 #include <iostream>
 
 void insert_sort(std::vector<int> &v) {
-    for (int i = 1; i < v.size(); i++) {
+    for (size_t i = 1; i < v.size(); i++) {
         int key = v[i];
         int j = i - 1;
         while (j >= 0 && key > v[j]) {
